requirements.txt
bcrypt
dill
faiss-cpu
google-generativeai
langchain==0.0.284
langchain_experimental
mysqlclient
mysql-connector-python
python-dotenv
PyMySQL
sqlalchemy
streamlit
